requirements.txt
pygame-ce==2.5.3
https://github.com/acemany/pyndustric/archive/refs/heads/master.zip
